Update MascotParser to merged Proline-Core 1.1.1 .
refs #7653

+++ pm-mascot-parser/src/main/scala/fr/proline/module/parser/mascot/MascotDataParser.scala
@@ -50,7 +50,7 @@ class MascotDataParser( val pepSummary: ms_peptidesummary,
    *
    *
    */
-  def parseMatches(rsId: Int): Boolean = {
+  def parseMatches(rsId: Long): Boolean = {
 
     //Get Necessary providers : for peptide, protein and ptms
     var pepProvider = parserContext.getProvider(classOf[IPeptideProvider])
